move dataset tags endpoints to dataset service
