+++ services/graph.js
@@ -31,12 +31,13 @@ export const fetchAllTags = (params = {}) => {
  */
 export const fetchInferredTags = (tags, params = {}) => {
   logger.info('Fetch inferred tags');
-  return WRIAPI.get(`graph/query/concepts-inferred?concepts=${tags}`,
+  return WRIAPI.get('graph/query/concepts-inferred',
     {
       headers: { 'Upgrade-Insecure-Requests': 1 },
       params: {
         env: process.env.API_ENV,
         application: process.env.APPLICATIONS,
+        concepts: tags,
         ...params
       }
     })
@@ -45,82 +46,6 @@ export const fetchInferredTags = (tags, params = {}) => {
       const { status, statusText } = response;
       logger.error(`Error fetching inferred tags: ${tags} ${status}: ${statusText}`);
       throw new Error(`Error inferred tags: ${tags} ${status}: ${statusText}`);
-    });
-};
-
-/**
-* Get dataset tags
-*/
-export const fetchDatasetTags = (datasetId, params = {}) => {
-  logger.info(`Fetch dataset tags: ${datasetId}`);
-  return WRIAPI.get(`dataset/${datasetId}/vocabulary`,
-    {
-      headers: { 'Upgrade-Insecure-Requests': 1 },
-      params: {
-        env: process.env.API_ENV,
-        application: process.env.APPLICATIONS,
-        ...params
-      }
-    })
-    .then(response => WRISerializer(response.data))
-    .catch((response) => {
-      const { status, statusText } = response;
-      logger.error(`Error fetching dataset tags ${datasetId}: ${status}: ${statusText}`);
-      throw new Error(`Error fetching dataset tags ${datasetId}: ${status}: ${statusText}`);
-    });
-};
-
-/**
-* Update dataset tags
-*/
-export const updateDatasetTags = (datasetId, tags, token, usePatch = false) => {
-  logger.info(`Update dataset tags: ${datasetId}`);
-
-  if (usePatch) {
-    return WRIAPI.patch(`dataset/${datasetId}/vocabulary/knowledge_graph`,
-      {
-        tags,
-        application: process.env.APPLICATIONS,
-        env: process.env.API_ENV
-      },
-      { headers: { Authorization: token } })
-      .then(response => WRISerializer(response.data))
-      .catch((response) => {
-        const { status, statusText } = response;
-        logger.error(`Error updating dataset tags ${datasetId}: ${status}: ${statusText}`);
-        throw new Error(`Error updating dataset tags ${datasetId}: ${status}: ${statusText}`);
-      });
-  }
-  if (tags.length > 0) {
-    return WRIAPI.put(`dataset/${datasetId}/vocabulary`,
-      {
-        knowledge_graph: {
-          tags,
-          application: process.env.APPLICATIONS,
-          env: process.env.API_ENV
-        }
-      },
-      { headers: { Authorization: token } })
-      .then(response => WRISerializer(response.data))
-      .catch((response) => {
-        const { status, statusText } = response;
-        logger.error(`Error updating dataset tags ${datasetId}: ${status}: ${statusText}`);
-        throw new Error(`Error updating dataset tags ${datasetId}: ${status}: ${statusText}`);
-      });
-  }
-  return WRIAPI.delete(`dataset/${datasetId}/vocabulary/knowledge_graph`,
-    {
-      headers: { Authorization: token },
-      params: {
-        application: process.env.APPLICATIONS,
-        env: process.env.API_ENV
-      }
-    })
-    .then(response => WRISerializer(response.data))
-    .catch((response) => {
-      const { status, statusText } = response;
-      logger.error(`Error updating dataset tags ${datasetId}: ${status}: ${statusText}`);
-      throw new Error(`Error updating dataset tags ${datasetId}: ${status}: ${statusText}`);
     });
 };
 
@@ -223,8 +148,6 @@ export default {
   fetchMostFavoritedDatasets,
   fetchSimilarDatasets,
   countDatasetView,
-  updateDatasetTags,
-  fetchDatasetTags,
   fetchInferredTags,
   fetchAllTags
 };
